Parser.cs splitting
