+++ src/DynamicExpresso.Core/Parsing/Types.cs
@@ -1,0 +1,287 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Dynamic;
+
+namespace DynamicExpresso.Parsing
+{
+	internal static class Types
+	{
+		public static bool IsNullableType(Type type)
+		{
+			return type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>);
+		}
+
+		public static bool IsDynamicType(Type type)
+		{
+			return typeof(IDynamicMetaObjectProvider).IsAssignableFrom(type);
+		}
+
+		private static Type GetNonNullableType(Type type)
+		{
+			return IsNullableType(type) ? type.GetGenericArguments()[0] : type;
+		}
+
+		public static string GetTypeName(Type type)
+		{
+			var baseType = GetNonNullableType(type);
+			var s = baseType.Name;
+			if (type != baseType) s += '?';
+			return s;
+		}
+
+		public static bool IsNumericType(Type type)
+		{
+			return GetNumericTypeKind(type) != 0;
+		}
+
+		private static bool IsSignedIntegralType(Type type)
+		{
+			return GetNumericTypeKind(type) == 2;
+		}
+
+		private static bool IsUnsignedIntegralType(Type type)
+		{
+			return GetNumericTypeKind(type) == 3;
+		}
+
+		private static int GetNumericTypeKind(Type type)
+		{
+			type = GetNonNullableType(type);
+			if (type.IsEnum) return 0;
+			switch (Type.GetTypeCode(type))
+			{
+				case TypeCode.Char:
+				case TypeCode.Single:
+				case TypeCode.Double:
+				case TypeCode.Decimal:
+					return 1;
+				case TypeCode.SByte:
+				case TypeCode.Int16:
+				case TypeCode.Int32:
+				case TypeCode.Int64:
+					return 2;
+				case TypeCode.Byte:
+				case TypeCode.UInt16:
+				case TypeCode.UInt32:
+				case TypeCode.UInt64:
+					return 3;
+				default:
+					return 0;
+			}
+		}
+
+		//static bool IsEnumType(Type type)
+		//{
+		//	return GetNonNullableType(type).IsEnum;
+		//}
+
+		public static bool IsCompatibleWith(Type source, Type target)
+		{
+			if (source == target)
+			{
+				return true;
+			}
+
+			if (!target.IsValueType)
+			{
+				return target.IsAssignableFrom(source);
+			}
+			var st = GetNonNullableType(source);
+			var tt = GetNonNullableType(target);
+			if (st != source && tt == target) return false;
+			var sc = st.IsEnum ? TypeCode.Object : Type.GetTypeCode(st);
+			var tc = tt.IsEnum ? TypeCode.Object : Type.GetTypeCode(tt);
+			switch (sc)
+			{
+				case TypeCode.SByte:
+					switch (tc)
+					{
+						case TypeCode.SByte:
+						case TypeCode.Int16:
+						case TypeCode.Int32:
+						case TypeCode.Int64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.Byte:
+					switch (tc)
+					{
+						case TypeCode.Byte:
+						case TypeCode.Int16:
+						case TypeCode.UInt16:
+						case TypeCode.Int32:
+						case TypeCode.UInt32:
+						case TypeCode.Int64:
+						case TypeCode.UInt64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.Int16:
+					switch (tc)
+					{
+						case TypeCode.Int16:
+						case TypeCode.Int32:
+						case TypeCode.Int64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.UInt16:
+					switch (tc)
+					{
+						case TypeCode.UInt16:
+						case TypeCode.Int32:
+						case TypeCode.UInt32:
+						case TypeCode.Int64:
+						case TypeCode.UInt64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.Int32:
+					switch (tc)
+					{
+						case TypeCode.Int32:
+						case TypeCode.Int64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.UInt32:
+					switch (tc)
+					{
+						case TypeCode.UInt32:
+						case TypeCode.Int64:
+						case TypeCode.UInt64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.Int64:
+					switch (tc)
+					{
+						case TypeCode.Int64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.UInt64:
+					switch (tc)
+					{
+						case TypeCode.UInt64:
+						case TypeCode.Single:
+						case TypeCode.Double:
+						case TypeCode.Decimal:
+							return true;
+					}
+					break;
+				case TypeCode.Single:
+					switch (tc)
+					{
+						case TypeCode.Single:
+						case TypeCode.Double:
+							return true;
+					}
+					break;
+				default:
+					if (st == tt) return true;
+					break;
+			}
+			return false;
+		}
+
+		// Return 1 if s -> t1 is a better conversion than s -> t2
+		// Return -1 if s -> t2 is a better conversion than s -> t1
+		// Return 0 if neither conversion is better
+		public static int CompareConversions(Type s, Type t1, Type t2)
+		{
+			if (t1 == t2) return 0;
+			if (s == t1) return 1;
+			if (s == t2) return -1;
+
+			var assignableT1 = t1.IsAssignableFrom(s);
+			var assignableT2 = t2.IsAssignableFrom(s);
+			if (assignableT1 && !assignableT2) return 1;
+			if (assignableT2 && !assignableT1) return -1;
+
+			var compatibleT1T2 = IsCompatibleWith(t1, t2);
+			var compatibleT2T1 = IsCompatibleWith(t2, t1);
+			if (compatibleT1T2 && !compatibleT2T1) return 1;
+			if (compatibleT2T1 && !compatibleT1T2) return -1;
+
+			if (IsSignedIntegralType(t1) && IsUnsignedIntegralType(t2)) return 1;
+			if (IsSignedIntegralType(t2) && IsUnsignedIntegralType(t1)) return -1;
+
+			return 0;
+		}
+
+		//static Type FindGenericType(Type generic, Type type)
+		//{
+		//	while (type != null && type != typeof(object))
+		//	{
+		//		if (type.IsGenericType && type.GetGenericTypeDefinition() == generic) return type;
+		//		if (generic.IsInterface)
+		//		{
+		//			foreach (Type intfType in type.GetInterfaces())
+		//			{
+		//				Type found = FindGenericType(generic, intfType);
+		//				if (found != null) return found;
+		//			}
+		//		}
+		//		type = type.BaseType;
+		//	}
+		//	return null;
+		//}
+		
+		public static IEnumerable<Type> SelfAndBaseTypes(Type type)
+		{
+			if (type.IsInterface)
+			{
+				var types = new List<Type>();
+				AddInterface(types, type);
+
+				types.Add(typeof(object));
+
+				return types;
+			}
+			return SelfAndBaseClasses(type);
+		}
+
+		private static IEnumerable<Type> SelfAndBaseClasses(Type type)
+		{
+			while (type != null)
+			{
+				yield return type;
+				type = type.BaseType;
+			}
+		}
+
+		private static void AddInterface(List<Type> types, Type type)
+		{
+			if (!types.Contains(type))
+			{
+				types.Add(type);
+				foreach (Type t in type.GetInterfaces())
+				{
+					AddInterface(types, t);
+				}
+			}
+		}
+	}
+}
